+++ src/compiler/expr/literal.zig
@@ -20,23 +20,23 @@ pub fn compileLiteral(state: *CompilerState, lit: *const ast.Literal) !void {
         .number => {
             if (std.mem.indexOfScalar(u8, lit.value, '.')) |_| {
                 const f = std.fmt.parseFloat(f64, lit.value) catch return error.CompileError;
-                try emit.emitConstant(state, .{ .float = f });
+                try emit.emitConstant(state, .{ .f64 = f });
             } else {
                 const n = std.fmt.parseInt(i64, lit.value, 10) catch return error.CompileError;
-                try emit.emitConstant(state, .{ .int = n });
+                try emit.emitConstant(state, .{ .i64 = n });
             }
         },
         .hex => {
             const n = std.fmt.parseInt(i64, lit.value[2..], 16) catch return error.CompileError;
-            try emit.emitConstant(state, .{ .int = n });
+            try emit.emitConstant(state, .{ .i64 = n });
         },
         .octal => {
             const n = std.fmt.parseInt(i64, lit.value[2..], 8) catch return error.CompileError;
-            try emit.emitConstant(state, .{ .int = n });
+            try emit.emitConstant(state, .{ .i64 = n });
         },
         .binary => {
             const n = std.fmt.parseInt(i64, lit.value[2..], 2) catch return error.CompileError;
-            try emit.emitConstant(state, .{ .int = n });
+            try emit.emitConstant(state, .{ .i64 = n });
         },
     }
 }
