+++ src/main.zig
@@ -223,7 +223,7 @@ fn smokeCmd(ctx: zli.CommandContext) !void {
     var c = llts.Chunk.init(ctx.allocator);
     defer c.deinit();
 
-    const idx = try c.addConstant(.{ .int = 42 });
+    const idx = try c.addConstant(.{ .i64 = 42 });
     try c.writeOp(.OP_CONSTANT);
     try c.write(@intCast((idx >> 8) & 0xff));
     try c.write(@intCast(idx & 0xff));
