+++ src/vm/builtins/log.zig
@@ -20,7 +20,7 @@ fn isErrorValue(vm: *VMState, v: Value) bool {
     const p = v.ptr;
     if (p < 1 or !vm.isValidHeapPtr(p - 1)) return false;
     const tag = vm.slot(p - 1).*;
-    return tag == .int and tag.int == ERROR_TAG;
+    return tag == .i64 and tag.i64 == ERROR_TAG;
 }
 
 /// Format an LLTS error for host logs (no redundant `Error:` prefix).
